Made extremely broad changes that can be summarised into new HTML , CSS, and Javascript files.
added features that affect the font size depending on the screen size,

the fixed header will not be in view when scrolling downward,

and so much more that can't be remembered at the moment.

i now see that i should have made my changes every single time i was done with a feature.